Client Liaison skeleton added to server
Client Liaison functions and variables added to main.cpp in ThreadPool.
A lot of placeholder types and such until we get more information. Most
things subject to change in the future. Still not settled on the base of
the liaison. Sure hope this doesn't blow everything up. This would have
been commited on monday if not for problems with my laptop.