This has yet to be merged with Geordie's GUI, as it's still in development. Basic functionality is there. The sheets inherently have an ID based on their position in the sheet, Ie; First 32x32 is 0, Second 32x32 is 1, etc..
The length/width of the spritesheet(in tiles) has to be manually entered into the main.as, but this works for testing purposes. Click on an area within the white square to select a tile, as well as its ID, and then draw on the appropriate area to get a visual representation as well as send the ID to that square.

Next up is collision map drawing, which won't be long at all, and then drawing entities. Which way be a little more work.